requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cycler==0.11.0
exceptiongroup==1.1.3
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gradio==3.34.0
gradio_client==0.2.6
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.5
linkify-it-py==2.0.2
llvmlite==0.39.1
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
oauth2client==4.1.3
onnxruntime==1.14.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orjson==3.9.7
outcome==1.2.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.0
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDrive==1.3.1
pydub==0.25.1
Pygments==2.17.2
PyMatting==1.1.12
pyparsing==3.1.1
pyreadline==2.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
rembg==2.0.40
requests==2.31.0
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
selenium==4.11.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.10.1
tifffile==2021.11.2
toolz==0.12.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==2.0.5
uvicorn==0.22.0
watchdog==3.0.0
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.3
zipp==3.15.0
